requirements.txt
pymorphy2
sentry-sdk
python-dateutil
requests